requirements.txt
mock
nose
python-statsd==1.6.0
